Added contents of C2000Ware v5.03.00.00

+++ device_support/f28003x/common/source/f28003x_cputimers.c
@@ -148,7 +148,7 @@ ConfigCpuTimer(struct CPUTIMER_VARS *Timer, float Freq, float Period)
     //
     Timer->CPUFreqInMHz = Freq;
     Timer->PeriodInUSec = Period;
-    temp = (long) (Freq * Period);
+    temp = (Uint32) (Freq * Period);
 
     //
     // Counter decrements PRD+1 times each period
