+++ device_support/f2803x/common/source/DSP2803x_CpuTimers.c
@@ -148,7 +148,7 @@ ConfigCpuTimer(struct CPUTIMER_VARS *Timer, float Freq, float Period)
     //
     Timer->CPUFreqInMHz = Freq;
     Timer->PeriodInUSec = Period;
-    PeriodInClocks = (long) (Freq * Period);
+    PeriodInClocks = (Uint32) (Freq * Period);
     
     //
     // Counter decrements PRD+1 times each period
