+++ device_support/f28p55x/common/source/f28p55x_cputimers.c
@@ -7,8 +7,10 @@
 //###########################################################################
 //
 //
-// $Copyright:
-// Copyright (C) 2022 Texas Instruments Incorporated - http://www.ti.com
+// 
+// C2000Ware v5.03.00.00
+//
+// Copyright (C) 2024 Texas Instruments Incorporated - http://www.ti.com
 //
 // Redistribution and use in source and binary forms, with or without 
 // modification, are permitted provided that the following conditions 
@@ -145,7 +147,7 @@ void ConfigCpuTimer(struct CPUTIMER_VARS *Timer, float Freq, float Period)
     //
     Timer->CPUFreqInMHz = Freq;
     Timer->PeriodInUSec = Period;
-    temp = (long) (Freq * Period);
+    temp = (Uint32) (Freq * Period);
 
     //
     // Counter decrements PRD+1 times each period
