+++ driverlib/f28004x/examples/ecap/ecap_ex2_capture_pwm.c
@@ -81,9 +81,9 @@
 uint32_t ecap1IntCount;
 uint32_t ecap1PassCount;
 uint32_t epwm3TimerDirection;
-volatile uint16_t cap2Count;
-volatile uint16_t cap3Count;
-volatile uint16_t cap4Count;
+volatile uint32_t cap2Count;
+volatile uint32_t cap3Count;
+volatile uint32_t cap4Count;
 volatile uint16_t epwm3PeriodCount;
 
 //
